% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{difference_significance_vitro}
\alias{difference_significance_vitro}
\title{difference_significance_vitro}
\usage{
difference_significance_vitro(
  norm_vitro,
  p_value = 0.1,
  marker_stages,
  marker_stages_filter,
  cluster_vitro,
  cluster_vivo,
  selected_stages,
  norm_vivo
)
}
\arguments{
\item{norm_vitro}{Norm count matrix (n_genes X n_cells) for in vitro dataset}

\item{p_value}{Numeric value.}

\item{marker_stages}{Second element of the list given as output by the function \emph{select_top_markers}}

\item{marker_stages_filter}{output from the function \emph{filter_in_vitro}}

\item{cluster_vitro}{cluster for in vitro dataset}

\item{cluster_vivo}{cluster for in vivo dataset}

\item{selected_stages}{In vivo stages for which the markers where computed with the function \emph{select_top_markers}}

\item{norm_vivo}{Norm count matrix (n_genes X n_cells) for in vivo dataset}
}
\value{
List with empirical p values
}
\description{
For each cluster, a centroid is defined as the mean expression value of \emph{marker_stages_filter} across cells in the cluster.
Comparison of in vitro and in vivo is done starting from a distance matrix based on spearman correlation between cluster centroid.
For a cluster A in the in vitro dataset, the Spearman’s correlation based distance between the cells and its centroid is computed. These values define an empirical distribution. The Spearman’s correlation based distance between the centroid of cluster A in the in vitro dataset and centroid of cluster B in the in vivo dataset is also computed.  An empirical p value is assigned to cluster B, from the comparison with the empirical distribution.
The distance between a pair of clusters is significant if the empirical p value above defined is smaller or equal to \emph{p_value}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
