% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_functions.R
\name{filter_in_vitro}
\alias{filter_in_vitro}
\title{filter_in_vitro}
\usage{
filter_in_vitro(
  norm_vitro,
  cluster_vitro,
  marker_all,
  fraction = 0.1,
  threshold = 0
)
}
\arguments{
\item{norm_vitro}{Norm count matrix (n_genes X n_cells) for in vitro dataset}

\item{cluster_vitro}{cluster for in vitro dataset}

\item{marker_all}{First element of the list given as output by the function \emph{select_top_markers}}

\item{fraction}{Numeric value.}

\item{threshold}{Numeric value}
}
\value{
Character vector with the names of kept genes
}
\description{
For a given gene in in \emph{marker_all}, if the fraction of cells in one or more clusters with an expression above \emph{threshold} is greater than \emph{fraction}, then the gene
is kept
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
