% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_published_methods.R
\name{findCellTypesScibet}
\alias{findCellTypesScibet}
\title{findCellTypesScibet}
\usage{
findCellTypesScibet(queryObj, referenceObj, namedLabels)
}
\arguments{
\item{queryObj}{Seurat object of the in vitro dataset}

\item{referenceObj}{Seurat object of the in vivo dataset}

\item{namedLabels}{refdata parameter in the function \emph{FindTransferAnchors}}
}
\value{
Seurat object
}
\description{
Projection done by function \emph{SciBet_R} from package \emph{scibetR}
}
\seealso{
\url{https://github.com/zwj-tina/scibetR}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
