% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_published_methods.R
\name{findCellTypesSeurat}
\alias{findCellTypesSeurat}
\title{findCellTypesSeurat}
\usage{
findCellTypesSeurat(
  queryObj,
  referenceObj,
  k.anchor = 5,
  k.filter = 200,
  namedLabels,
  k.weight = 50
)
}
\arguments{
\item{queryObj}{Seurat object of the in vitro dataset}

\item{referenceObj}{Seurat object of the in vivo dataset}

\item{k.anchor}{k.anchor parameter in the function \emph{FindTransferAnchors}}

\item{k.filter}{k.filter parameter in the function \emph{FindTransferAnchors}}

\item{namedLabels}{refdata parameter in the function \emph{FindTransferAnchors}}

\item{k.weight}{k.weight parameter in the function \emph{TransferData}}
}
\value{
Seurat object
}
\description{
Projection done by functions \emph{FindTransferAnchors} and \emph{TransferData} from package \emph{Seurat}
}
\seealso{
\url{https://CRAN.R-project.org/package=Seurat}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
