% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{heatmap_markers_vitro}
\alias{heatmap_markers_vitro}
\title{heatmap_markers_vitro}
\usage{
heatmap_markers_vitro(
  norm_vitro,
  marker_plot,
  cluster_vitro,
  color_vector = NULL
)
}
\arguments{
\item{norm_vitro}{Norm count matrix (n_genes X n_cells) for in vitro dataset}

\item{marker_plot}{Character vector with the names of the genes to plot.}

\item{cluster_vitro}{cluster for in vitro dataset}

\item{color_vector}{Character vector with colour assignment.}
}
\value{
Heatmap class object
}
\description{
heatmap_markers_vitro
}
\seealso{
\url{https://www.rdocumentation.org/packages/ComplexHeatmap/versions/1.10.2/topics/Heatmap}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
