% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{heatmap_markers_vivo_median_profile}
\alias{heatmap_markers_vivo_median_profile}
\title{heatmap_markers_vivo_median_profile}
\usage{
heatmap_markers_vivo_median_profile(
  norm_vivo,
  marker_stages,
  selected_stages,
  max_number = 10,
  cluster_vivo,
  color_vector = NULL
)
}
\arguments{
\item{norm_vivo}{Norm count matrix (n_genes X n_cells) for in vivo dataset}

\item{marker_stages}{Second element of the list given as output by the function \emph{select_top_markers}}

\item{selected_stages}{In vivo stages for which the markers where computed with the function \emph{select_top_markers}}

\item{max_number}{Integer. Maximum number of genes to show for each \emph{selected_stages}.}

\item{cluster_vivo}{cluster for in vivo dataset}

\item{color_vector}{Character vector with colour assignment.}
}
\value{
Heatmap class object
}
\description{
heatmap_markers_vivo_median_profile
}
\seealso{
\url{https://www.rdocumentation.org/packages/ComplexHeatmap/versions/1.10.2/topics/Heatmap}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
