% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_barplot}
\alias{plot_barplot}
\title{plot_barplot}
\usage{
plot_barplot(obj, method, cluster_vitro, threshold, fraction = TRUE)
}
\arguments{
\item{obj}{Seurat object given as output from functions \emph{findCellTypesSeurat}, \emph{findCellTypesScibet} or
\emph{findCellTypes_scmap}}

\item{method}{Character name. Must be one of Seurat, scibetR or scmap}

\item{cluster_vitro}{cluster for in vitro dataset}

\item{threshold}{Numeric value. For a given gene, the jaccard index between the links from the in vivo and in vitro datasets is computed. If the jaccard index is above \emph{threshold}, then the gene is considered to be conserved between the two datasets.}

\item{fraction}{Logical. If TRUE, for each \emph{cluster_vitro}, the fraction of cells
mapped to the in vivo stages are shown. If FALSE, the number of cells mapped to the in vivo stages are shown}
}
\value{
ggplot2 object.
}
\description{
plot_barplot
}
\seealso{
\url{https://CRAN.R-project.org/package=ggplot2}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
