% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{plot_boxplot}
\usage{
plot_boxplot(
  norm_vivo,
  gene_name,
  cluster_vivo,
  order_label_vivo,
  title_name,
  fill_name,
  x_name,
  y_name
)
}
\arguments{
\item{norm_vivo}{Norm count matrix (n_genes X n_cells)}

\item{gene_name}{Character name of the gene of interest}

\item{cluster_vivo}{cluster partition}

\item{order_label_vivo}{Character vector specifyng the order of the columns in the balloon plot. The names must be present in \emph{cluster_vivo_factor}}

\item{title_name}{Character value.}

\item{fill_name}{Character value.}

\item{x_name}{Character name of x axis.}

\item{y_name}{Character value}
}
\value{
ggplot2 object.
}
\description{
Only the stages in \emph{cluster_vivo_factor} with an assignment greater than zero are showns in the balloon plot
}
\seealso{
\url{https://CRAN.R-project.org/package=gplots}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
