% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_distance_vivo_vitro}
\alias{plot_distance_vivo_vitro}
\title{plot_distance_vivo_vitro}
\usage{
plot_distance_vivo_vitro(
  norm_vivo,
  cluster_vivo,
  selected_stages,
  marker_stages,
  marker_stages_filter,
  norm_vitro,
  cluster_vitro,
  text_size = 10,
  show_significance = FALSE,
  p_value = 0.1
)
}
\arguments{
\item{norm_vivo}{Norm count matrix (n_genes X n_cells) for in vivo dataset}

\item{cluster_vivo}{cluster for in vivo dataset}

\item{selected_stages}{In vivo stages for which the markers where computed with the function \emph{select_top_markers}}

\item{marker_stages}{Second element of the list given as output by the function \emph{select_top_markers}}

\item{marker_stages_filter}{output from the function \emph{filter_in_vitro}}

\item{norm_vitro}{Norm count matrix (n_genes X n_cells) for in vitro dataset}

\item{cluster_vitro}{cluster for in vitro dataset}

\item{text_size}{Integer number specifying the size of the text.}

\item{show_significance}{Logical. If TRUE the function \emph{difference_significance_vitro} is run and an asterisk is shown if the empirical p value is below \emph{p_value}}

\item{p_value}{Numeric value.}
}
\value{
Heatmap class object
}
\description{
plot_distance_vivo_vitro
}
\seealso{
\url{https://www.rdocumentation.org/packages/ComplexHeatmap/versions/1.10.2/topics/Heatmap}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
