% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_in_vivo_markers}
\alias{plot_in_vivo_markers}
\title{plot_in_vivo_markers}
\usage{
plot_in_vivo_markers(
  coordinate_umap,
  norm_vitro,
  marker_stages,
  selected_stages,
  name_vivo,
  threshold = 0,
  title_name
)
}
\arguments{
\item{coordinate_umap}{data frame with umap coordinates (UMAP 1 and UMAP2).}

\item{norm_vitro}{Norm count matrix (n_genes X n_cells) for in vitro dataset}

\item{marker_stages}{Second element of the list given as output by the function \emph{select_top_markers}}

\item{selected_stages}{In vivo stages for which the markers where computed with the function \emph{select_top_markers}}

\item{name_vivo}{name of the in vivo stage on which SCOPRO is run}

\item{threshold}{Numeric value.}

\item{title_name}{Character value.}
}
\value{
ggplot2::ggplot2 object.
}
\description{
Cells are coloured according to the number of markers of the in vivo stage \emph{name_vivo} that are epxressed above \emph{threshold}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
