% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_in_vivo_markers_interactive}
\alias{plot_in_vivo_markers_interactive}
\title{plot_in_vivo_markers_interactive}
\usage{
plot_in_vivo_markers_interactive(
  coordinate_umap,
  norm_vitro,
  marker_stages,
  selected_stages,
  name_vivo,
  threshold = 0,
  max_number = 5,
  min_x = NULL,
  max_x = NULL,
  min_y = NULL,
  max_y = NULL
)
}
\arguments{
\item{coordinate_umap}{data frame with umap coordinates (UMAP 1 and UMAP2).}

\item{norm_vitro}{Norm count matrix (n_genes X n_cells) for in vitro dataset}

\item{marker_stages}{Second element of the list given as output by the function \emph{select_top_markers}}

\item{selected_stages}{In vivo stages for which the markers where computed with the function \emph{select_top_markers}}

\item{name_vivo}{name of the in vivo stage on which SCOPRO is run}

\item{threshold}{Numeric value. For a given gene, the jaccard index between the links from the in vivo and in vitro datasets is computed. If the jaccard index is above \emph{threshold}, then the gene is considered to be conserved between the two datasets.}

\item{max_number}{Integer. Maximum number of genes name to show.}

\item{min_x}{Set the min limit on the x axis.}

\item{max_x}{Set the max limit on the x axis.}

\item{min_y}{Set the min limit on the y axis.}

\item{max_y}{Set the min limit on the y axis.}
}
\value{
plotly object given by \emph{plot_ly function} (from library \emph{plotly}).
}
\description{
Cells are coloured according to the number of markers of the in vivo stage \emph{name_vivo} that are expressed above \emph{threshold}.
It is based on plotly library.
}
\seealso{
\url{https://plotly.com/r/}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
