% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_score}
\alias{plot_score}
\title{plot_score}
\usage{
plot_score(
  SCOPRO_output,
  marker_stages,
  marker_stages_filter,
  selected_stages,
  name_vivo,
  y_name,
  fill_name,
  title_name
)
}
\arguments{
\item{SCOPRO_output}{output given by function \emph{SCOPRO}}

\item{marker_stages}{Second element of the list given as output by the function \emph{select_top_markers}}

\item{marker_stages_filter}{output from the function \emph{filter_in_vitro}}

\item{selected_stages}{In vivo stages for which the markers where computed with the function \emph{select_top_markers}}

\item{name_vivo}{name of the in vivo stage on which SCOPRO is run}

\item{y_name}{Character value}

\item{fill_name}{Character value.}

\item{title_name}{Character value.}
}
\value{
ggplot2::ggplot2 object.
}
\description{
plot_score
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
