% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_score_genes}
\alias{plot_score_genes}
\title{plot_score_genes}
\usage{
plot_score_genes(
  markers_to_plot,
  label_1,
  label_2,
  norm_vitro,
  norm_vivo,
  cluster_vitro,
  cluster_vivo,
  final_name,
  max_size = 9,
  text_size = 9.5,
  title_name,
  norm_expression = FALSE
)
}
\arguments{
\item{markers_to_plot}{Character vector with the names of the genes to plot.}

\item{label_1}{Character value. Label for the in vitro dataset}

\item{label_2}{Character value. Label for the in vivo dataset}

\item{norm_vitro}{Norm count matrix (n_genes X n_cells) for in vitro dataset}

\item{norm_vivo}{Norm count matrix (n_genes X n_cells) for in vivo dataset}

\item{cluster_vitro}{cluster for in vitro dataset}

\item{cluster_vivo}{cluster for in vivo dataset}

\item{final_name}{Character vector with the names of the genes to show in the plot.}

\item{max_size}{Numeric value, specifying the size of the dot.}

\item{text_size}{Numeric value, specifying the size of the text in the plot.}

\item{title_name}{Character value.}

\item{norm_expression}{Logical. If TRUE, for a given gene, the average expression is computed in each cluster. Finally, for each cluster, the average expression of the genes are normalised (divide by the maximum average expression). The color of each dot is the normalised average expression.
If FALSE, the color is the average expression of the gene in the corresponding cluster.}
}
\value{
ggplot2::ggplot2 object showing balloon plot.The size of the dot is given by the fraction of cells with log2 normalise expression above 0.1
}
\description{
plot_score_genes
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
