% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{SCOPRO_random}
\alias{SCOPRO_random}
\title{SCOPRO_random}
\usage{
SCOPRO_random(
  norm_vitro,
  norm_vivo,
  cluster_vitro,
  cluster_vivo,
  name_vivo,
  marker_stages_filter,
  fold_change = 3,
  threshold_fold_change = 0.1,
  selected_stages,
  iteration = 100
)
}
\arguments{
\item{norm_vitro}{Norm count matrix (n_genes X n_cells) for in vitro dataset}

\item{norm_vivo}{Norm count matrix (n_genes X n_cells) for in vivo dataset}

\item{cluster_vitro}{cluster for in vitro dataset}

\item{cluster_vivo}{cluster for in vivo dataset}

\item{name_vivo}{name of the in vivo stage on which SCOPRO is run}

\item{marker_stages_filter}{output from the function \emph{filter_in_vitro}}

\item{fold_change}{Numeric value. For a given gene, the fold change between all the other genes is computed. If fold change is above \emph{fold_change}, then there is a link with weight 1 between the two genes.}

\item{threshold_fold_change}{Numeric value. Above \emph{threshold} the fold change between genes is computed. Below \emph{threshold} the difference between genes is computed.}

\item{selected_stages}{In vivo stages for which the markers where computed with the function \emph{select_top_markers}}

\item{iteration}{Integer. Number of time SCOPRO algorithm should run starting from}
}
\value{
A list with length equal to the number of \emph{cluster_vivo}. Each entry is a vetor with length equal to \emph{iteration} with the fraction of conserved links between the \emph{name_vivo} network and the network
from each cluster in the in vitro dataset.
}
\description{
Run SCOPRO function for a number of times equal to \emph{iteration}. Each time the original links in the in vivo and in vitro networks are randomly
reassigned.
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
