% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_functions.R
\name{select_no_common_genes_sc}
\alias{select_no_common_genes_sc}
\title{select_no_common_genes_sc}
\usage{
select_no_common_genes_sc(
  SCOPRO_output,
  marker_stages,
  selected_stages,
  name_vivo,
  cluster_vitro,
  name_vitro,
  threshold = 0.5
)
}
\arguments{
\item{SCOPRO_output}{output given by function \emph{SCOPRO_single_cell}}

\item{marker_stages}{Second element of the list given as output by the function \emph{select_top_markers}}

\item{selected_stages}{In vivo stages for which the markers where computed with the function \emph{select_top_markers}}

\item{name_vivo}{name of the in vivo stage on which SCOPRO is run}

\item{cluster_vitro}{cluster for in vitro dataset}

\item{name_vitro}{name of the in vitro stage for which we want to know the non-conserved markers with the \emph{name_vivo} stage}

\item{threshold}{fraction of cells in the in vitro cluster \emph{name_vitro}.
Only the markers not conserved in more than \emph{threshold} fraction of cells are given as output.}
}
\value{
Character vector with the names of the non-conserved markers of \emph{name_vivo} stage in the \emph{name_vitro} cluster
}
\description{
select_no_common_genes_sc
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
