% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{significance_value}
\alias{significance_value}
\title{significance_value}
\usage{
significance_value(SCOPRO_output, SCOPRO_output_random, threshold = 0.1)
}
\arguments{
\item{SCOPRO_output}{output given by function \emph{SCOPRO}}

\item{SCOPRO_output_random}{Output given by the function \emph{SCOPRO_random}.}

\item{threshold}{Numeric value.}
}
\description{
Detect the in vitro clusters for which the difference between the final score given by SCOPRO and the median value of the final scores given by the function SCOPRO_random
is greater than \emph{threshold}.
}
