% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{cellTypesPerClusterBalloonPlot_small}
\alias{cellTypesPerClusterBalloonPlot_small}
\title{cellTypesPerClusterBalloonPlot_small}
\usage{
cellTypesPerClusterBalloonPlot_small(
  obj,
  cluster_vivo_factor,
  title_name,
  text_size = 0.4,
  label_size = 0.7,
  thresold = 0.5
)
}
\arguments{
\item{obj}{Seurat object given as output from function \emph{findCellTypesSeurat}}

\item{cluster_vivo_factor}{Factor vector specifyng the cluster partition of the in vivo datataset}

\item{title_name}{Character value.}

\item{text_size}{Numeric value, specifying the size of the text in the plot.}

\item{label_size}{Numeric value specifyng the label.size parameter in the function \emph{balloonplot} from package \emph{gplots}.}

\item{thresold}{Numeric value. Cells with a predicted score below or equal to \emph{thresold} are labelled as unassigned}
}
\value{
balloon plot given by function \emph{balloonplot} from package \emph{gplots}
}
\description{
Only the stages in \emph{cluster_vivo_factor} with an assignment greater than zero are showns in the balloon plot
}
\seealso{
\url{https://CRAN.R-project.org/package=gplots}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
