% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{cellTypesPerClusterBalloonPlot_small}
\alias{cellTypesPerClusterBalloonPlot_small}
\title{cellTypesPerClusterBalloonPlot_small}
\usage{
cellTypesPerClusterBalloonPlot_small(
  obj,
  cluster_vivo_factor,
  order_label_vivo,
  title_name,
  method = "Seurat",
  text_size = 0.4,
  label_size = 0.7,
  threshold = 0.5
)
}
\arguments{
\item{obj}{Seurat object given as output from function \emph{findCellTypesSeurat}}

\item{cluster_vivo_factor}{Factor vector specifying the cluster partition of the in vivo datataset}

\item{order_label_vivo}{Character vector specifying the order of the columns in the balloon plot. The names must be present in \emph{cluster_vivo_factor}}

\item{title_name}{Character value.}

\item{method}{Character name. Must be one of Seurat, scibetR or scmap}

\item{text_size}{Numeric value, specifying the size of the text in the plot.}

\item{label_size}{Numeric value specifying the label.size parameter in the function \emph{balloonplot} from package \emph{gplots}.}

\item{threshold}{Numeric value. For a given gene, the jaccard index between the links from the in vivo and in vitro datasets is computed. If the jaccard index is above \emph{threshold}, then the gene is considered to be conserved between the two datasets.}
}
\value{
balloon plot given by function \emph{balloonplot} from package \emph{gplots}
}
\description{
Only the stages in \emph{cluster_vivo_factor} with an assignment greater than zero are shown the balloon plot
}
\seealso{
\url{https://CRAN.R-project.org/package=gplots}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
