% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{significance_value}
\alias{significance_value}
\title{significance_value}
\usage{
significance_value(SCOPRO_output, threshold = 0.05)
}
\arguments{
\item{SCOPRO_output}{output given by function \emph{SCOPRO}}

\item{threshold}{Numeric value.}
}
\description{
Detect the in vitro clusters for which the final score given by SCOPRO is statistically significant
(p_value equal or below than \emph{threshold}), under the assumption of a binomial distribution.
is greater than \emph{threshold}.
}
